# Backend/auth-check/openapi.yaml
openapi: 3.0.3
info:
  title: AutoCheck API
  version: v1
  description: |
    REST API платформы автоматической проверки тестовых заданий разработчиков.

    ## Роли
    - **EXPERT** — создаёт задания, видит все проверки, выносит вердикт
    - **CANDIDATE** — загружает решение, видит только свои проверки

    ## Авторизация
    Все эндпоинты кроме `/auth/login` и `/auth/register` требуют заголовок:
    ```
    Authorization: Bearer <token>
    ```

    ## Формат ответов
    Все ответы завёрнуты в единый конверт `{data, error, meta}`.
    Null-поля **не включаются** в ответ (NON_NULL Jackson).
    - **Успех**: `{"data": {...}}`
    - **Ошибка**: `{"error": {"code": "...", "message": "..."}}`
servers:
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
...
paths:
...
  /candidates:
    get:
      tags: [Candidates]
      summary: Список всех кандидатов (только EXPERT)
      responses:
        '200':
          description: Список кандидатов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateListEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /candidates/{id}:
    parameters:
      - $ref: '#/components/parameters/IdPath'
    get:
      tags: [Candidates]
      summary: Кандидат по ID (только EXPERT)
      responses:
        '200':
          description: Кандидат
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CandidateEnvelope'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Reports ───────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  # ── Reusable responses ────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Токен отсутствует, истёк или отозван
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: UNAUTHORIZED
              message: Full authentication is required to access this resource
    Forbidden:
      description: Недостаточно прав (требуется роль EXPERT)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: FORBIDDEN
              message: Нет прав доступа
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
          example:
            error:
              code: NOT_FOUND
              message: Задание 99 не найдено
...
  schemas:
...
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: NOT_FOUND
        message:
          type: string
          example: Задание 5 не найдено
        details:
          type: object
          additionalProperties:
            type: string
          description: Ошибки по полям — только при 422
    ErrorEnvelope:
      type: object
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
        fullName:
          type: string
        submissionsCount:
          type: integer
        bestScore:
          type: number
          format: double
          description: Лучший балл среди завершённых проверок. Отсутствует если проверок нет.
    CandidateEnvelope:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Candidate'
    CandidateListEnvelope:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'

    # ── Reports ───────────────────────────────────────────────────────────────
...